gometalinter: run on subpkgs too (#878)

+++ blockchain/difficulty.go
@@ -767,7 +767,7 @@ func estimateSupply(params *chaincfg.Params, height int64) int64 {
 	// interval then adding the subsidy produced by number of blocks in the
 	// current interval.
 	supply := params.BlockOneSubsidy()
-	reductions := int64(height) / params.SubsidyReductionInterval
+	reductions := height / params.SubsidyReductionInterval
 	subsidy := params.BaseSubsidy
 	for i := int64(0); i < reductions; i++ {
 		supply += params.SubsidyReductionInterval * subsidy
@@ -775,7 +775,7 @@ func estimateSupply(params *chaincfg.Params, height int64) int64 {
 		subsidy *= params.MulSubsidy
 		subsidy /= params.DivSubsidy
 	}
-	supply += (1 + int64(height)%params.SubsidyReductionInterval) * subsidy
+	supply += (1 + height%params.SubsidyReductionInterval) * subsidy
 
 	// Blocks 0 and 1 have special subsidy amounts that have already been
 	// added above, so remove what their subsidies would have normally been
@@ -1440,7 +1440,7 @@ func (b *BlockChain) estimateNextStakeDifficultyV2(curNode *blockNode, newTicket
 
 	// Calculate the number of votes that will occur during the remainder of
 	// the interval.
-	stakeValidationHeight := int64(b.chainParams.StakeValidationHeight)
+	stakeValidationHeight := b.chainParams.StakeValidationHeight
 	var pendingVotes int64
 	if nextRetargetHeight > stakeValidationHeight {
 		votingBlocks := blocksUntilRetarget - 1
